production

+++ Application/Vehicles/Commands/UpdateVehicle/UpdateVehicleCommand.cs
@@ -11,9 +11,9 @@ public class UpdateVehicleCommand : IRequest
     public DateOnly? FirstRegistrationDate { get; set; }
     public DateOnly? ProductionYear { get; set; }
     public string? VIN { get; set; }
-    public decimal? KW { get; set; }
-    public decimal? KM { get; set; }
-    public decimal? Capacity { get; set; }
+    public int? KW { get; set; }
+    public int? KM { get; set; }
+    public int? Capacity { get; set; }
     public uint? Mileage { get; set; }
 
     public Guid InsurerId { get; set; }
