+++ Application/Vehicles/Dtos/VehicleDto.cs
@@ -10,9 +10,9 @@ public class VehicleDto
     public DateOnly? FirstRegistrationDate { get; set; }
     public DateOnly? ProductionYear { get; set; }
     public string? VIN { get; set; }
-    public decimal? KW { get; set; }
-    public decimal? KM { get; set; }
-    public decimal? Capacity { get; set; }
+    public int? KW { get; set; }
+    public int? KM { get; set; }
+    public int? Capacity { get; set; }
     public uint? Mileage { get; set; }
     public VehicleBrandDto VehicleBrand { get; set; } = null!;
