+++ PolisProReminder/Requests/CreateVehicleReq.cs
@@ -7,9 +7,9 @@ public class CreateVehicleReq
     public string FirstRegistrationDate { get; set; }
     public string ProductionYear { get; set; }
     public string? VIN { get; set; }
-    public decimal? KW { get; set; }
-    public decimal? KM { get; set; }
-    public decimal? Capacity { get; set; }
+    public int? KW { get; set; }
+    public int? KM { get; set; }
+    public int? Capacity { get; set; }
     public uint? Mileage { get; set; }
 
     public string InsurerId { get; set; }
